+++ src/op/conv_zh.rs
@@ -6,8 +6,8 @@ use opencc_rust::{DefaultConfig, generate_static_dictionary, OpenCC};
 
 use crate::model::{MyTag, TEXT_TAGS};
 use crate::model::ConvZhProfile;
-use crate::op::{get_file_iterator, get_tags_from_args, get_where, string_to_option, WriteTextForCurrentAction};
-use crate::op::{Action, WalkAction, WriteAction, WriteTextAction};
+use crate::op::{get_file_iterator, get_tags_from_args, get_where, string_to_option};
+use crate::op::{Action, WalkAction, WriteAction, WriteTextAction, WriteTextForCurrentAction};
 use crate::op::tag_impl::ReadWriteTag;
 use crate::where_clause::WhereClause;