+++ src/op/ren.rs
@@ -4,9 +4,10 @@ use std::path::{Path, PathBuf};
 use anyhow::{anyhow, Error};
 use log::{error, info};
 
-use crate::model::MyTag;
-use crate::op::{get_file_iterator, get_tags_from_template, get_where, MyValues};
-use crate::op::{Action, ReadAction, WalkAction};
+use crate::model::{FilenameExistPolicy, MyTag};
+use crate::op::{get_file_iterator, get_new_path, get_tags_from_template, get_where};
+use crate::op::{Action, MyValues, ReadAction, WalkAction};
+use crate::util::path::combine_path;
 use crate::where_clause::WhereClause;
 
 pub struct RenAction {
@@ -16,6 +17,7 @@ pub struct RenAction {
     where_clause: Option<WhereClause>,
     with_properties: bool,
     template: String,
+    filename_exist_policy: FilenameExistPolicy,
     // for cache
     empty_value: String,
 }
@@ -24,7 +26,8 @@ impl RenAction {
     pub fn new<P>(dir: P,
                   dry_run: bool,
                   where_string: &Option<String>,
-                  template: &str) -> Result<Self, Error>
+                  template: &str,
+                  filename_exist_policy: FilenameExistPolicy) -> Result<Self, Error>
         where P: AsRef<Path>
     {
         let it = get_file_iterator(dir.as_ref())?;
@@ -38,6 +41,7 @@ impl RenAction {
             where_clause,
             with_properties: false,
             template: template.to_owned(),
+            filename_exist_policy,
             empty_value,
         })
     }
@@ -64,36 +68,6 @@ impl WalkAction for RenAction {
 
     fn tags(&self) -> &Vec<MyTag> {
         &self.tags
-    }
-}
-
-fn get_new_path<P>(path: P,
-                   stem: &str) -> Option<PathBuf>
-    where P: AsRef<Path>
-{
-    let old_path = path.as_ref();
-    let mut i: u16 = 0;
-    loop {
-        let mut filename = stem.to_owned();
-        if i > 0 {
-            filename.push_str(&format!("({})", i));
-        }
-        if let Some(ext) = old_path.extension() {
-            filename.push('.');
-            filename.push_str(&ext.to_string_lossy());
-        }
-        let mut new_path = PathBuf::from(old_path);
-        new_path.set_file_name(&filename);
-
-        if !new_path.exists() {
-            break Some(new_path);
-        } else {
-            if i == u16::MAX {
-                error!("Could NOT get the new filename! old_path: {:?}, stem: {}",  old_path, stem);
-                break None;
-            }
-            i = i + 1;
-        }
     }
 }
 
@@ -124,23 +98,24 @@ impl ReadAction for RenAction {
     }
 
     fn do_output(&mut self, path: &Path, content: &str) -> Result<bool, Error> {
-        if let Some(ref new_path) = get_new_path(path, &content) {
-            if !self.dry_run {
-                fs::rename(path, new_path).map_or_else(
-                    |e| {
-                        Err(anyhow!("Rename file {:?} to {:?} failed. (error: {:?})",
-                            path, new_path, e))
-                    },
-                    |_| {
-                        info!("Rename file {:?} to {:?}", path, new_path);
-                        Ok(true)
-                    })
-            } else {
-                ren_dry_run(path, new_path)
+        if let Some(p) = combine_path(path, content) {
+            if path.eq(&p) {
+                return Ok(true);
             }
-        } else {
-            Ok(false)
+
+            if let Some(ref new_path) = get_new_path(&p, self.filename_exist_policy) {
+                return if !self.dry_run {
+                    ren_real(self.filename_exist_policy, path, new_path)
+                } else {
+                    ren_dry_run(self.filename_exist_policy, path, new_path)
+                };
+            }
         }
+
+        error!(
+            "Could NOT get the new filename! path: {:?}, content: {}, filename_exist_policy: {:?}",
+            path, & content, self.filename_exist_policy);
+        Ok(false)
     }
 }
 
@@ -153,15 +128,50 @@ fn get_empty_value(tags: &[MyTag], template: &str) -> String {
     result
 }
 
-fn ren_dry_run<P>(path: P, new_path: P) -> Result<bool, Error>
-    where P: AsRef<Path>
-{
-    let new_path = new_path.as_ref();
+fn ren_real<P1, P2>(filename_exist_policy: FilenameExistPolicy,
+                    path: P1,
+                    new_path: P2) -> Result<bool, Error>
+    where P1: AsRef<Path>,
+          P2: AsRef<Path> {
     let path = path.as_ref();
+    let new_path = new_path.as_ref();
+    match filename_exist_policy {
+        FilenameExistPolicy::Overwrite => {
+            if new_path.exists() && new_path.is_file() {
+                let _ = fs::remove_file(new_path);
+            }
+        }
+        _ => {}
+    }
+    fs::rename(path, new_path).map_or_else(
+        |e| {
+            Err(anyhow!("Rename file {:?} to {:?} failed. (error: {:?})",
+                            path, new_path, e))
+        },
+        |_| {
+            info!("Rename file {:?} to {:?}", path, new_path);
+            Ok(true)
+        })
+}
+
+fn ren_dry_run<P1, P2>(filename_exist_policy: FilenameExistPolicy,
+                       path: P1,
+                       new_path: P2) -> Result<bool, Error>
+    where P1: AsRef<Path>,
+          P2: AsRef<Path> {
+    let path = path.as_ref();
+    let new_path = new_path.as_ref();
     if !new_path.exists() {
         info!("Rename file {:?} to {:?}", path, new_path);
         Ok(true)
     } else {
-        Err(anyhow!("Rename file {:?} to {:?} failed! File already exists.", path, new_path))
+        match filename_exist_policy {
+            FilenameExistPolicy::Overwrite if new_path.exists() && new_path.is_file() => {
+                info!("Rename file {:?} to {:?}", path, new_path);
+                Ok(true)
+            }
+            _ => Err(anyhow!("Rename file {:?} to {:?} failed! File already exists.",
+                path, new_path))
+        }
     }
 }