+++ src/util/mod.rs
@@ -1,3 +1,4 @@
 pub mod json_de;
 pub mod numeric;
+pub mod path;
 pub mod str;